Refactor Code reorganization
