+++ app/src/main/java/com/godamy/myplayer/ui/detail/DetailUiState.kt
@@ -1,6 +1,6 @@
 package com.godamy.myplayer.ui.detail
 
-import com.godamy.myplayer.data.Error
+import com.godamy.myplayer.domain.Error
 import com.godamy.myplayer.domain.MediaItem
 
 data class DetailUiState(
