+++ app/src/main/java/com/godamy/myplayer/ui/main/MainFragment.kt
@@ -11,9 +11,9 @@ import com.godamy.myplayer.data.MediaRepository
 import com.godamy.myplayer.data.RegionRepository
 import com.godamy.myplayer.databinding.FragmentMainBinding
 import com.godamy.myplayer.framework.AndroidPermissionChecker
-import com.godamy.myplayer.framework.datasource.MediaItemRoomDataSource
-import com.godamy.myplayer.framework.datasource.MediaItemServerDataSource
-import com.godamy.myplayer.framework.datasource.PlayServiceLocationDataSource
+import com.godamy.myplayer.framework.PlayServiceLocationDataSource
+import com.godamy.myplayer.framework.database.MediaItemRoomDataSource
+import com.godamy.myplayer.framework.server.MediaItemServerDataSource
 import com.godamy.myplayer.ui.main.adapter.MediaItemAdapter
 import com.godamy.myplayer.usecases.GetPopularMoviesUserCase
 import com.godamy.myplayer.usecases.RequestPopularMoviesUseCase
