+++ app/src/main/java/com/godamy/myplayer/usecases/RequestPopularMoviesUseCase.kt
@@ -1,7 +1,7 @@
 package com.godamy.myplayer.usecases
 
-import com.godamy.myplayer.data.Error
 import com.godamy.myplayer.data.MediaRepository
+import com.godamy.myplayer.domain.Error
 
 class RequestPopularMoviesUseCase(private val repository: MediaRepository) {
     suspend operator fun invoke(): Error? = repository.requestPopularMovies()
