+++ app/src/main/java/com/godamy/myplayer/usecases/SwitchFavoriteUseCase.kt
@@ -1,7 +1,7 @@
 package com.godamy.myplayer.usecases
 
-import com.godamy.myplayer.data.Error
 import com.godamy.myplayer.data.MediaRepository
+import com.godamy.myplayer.domain.Error
 import com.godamy.myplayer.domain.MediaItem
 
 class SwitchFavoriteUseCase(private val repository: MediaRepository) {
